requirements.txt
aniso8601==9.0.1
APScheduler==3.7.0
attrs==21.2.0
autopep8==1.6.0
certifi==2021.5.30
charset-normalizer==2.0.4
click==8.0.1
Flask==1.1.2
Flask-Cors==3.0.10
flask-restplus==0.13.0
idna==3.2
importlib-metadata==4.6.3
itsdangerous==2.0.1
Jinja2==3.0.1
jsonschema==3.2.0
MarkupSafe==2.0.1
Pillow==8.3.1
psycopg2-binary==2.8.6
pycodestyle==2.8.0
PyJWT==1.4.2
pyrsistent==0.18.0
pytesseract==0.3.8
pytz==2021.1
PyYAML==5.4.1
requests==2.26.0
six==1.16.0
SQLAlchemy==1.3.23
toml==0.10.2
typing-extensions==3.10.0.0
tzlocal==2.1
urllib3==1.26.6
Werkzeug==0.16.1
zipp==3.5.0
